Add configuration specification
This adds the initial specification for all configuration options. It's
only a first prototype and not complete.

This lays out an abstract grammar which can be represented as JSON/YAML
but is not any valid configuration syntax. It closely resembles Rust to
make sure that a representation in Rust code does not cause any issues.