Add config deserialization structs
The struct version of the config format specification has now been
implemented. This is a complete Rust representation of the config using
`serde` to (de)serialize the structure.

Some basic tests have also been added to make sure that the
deserialization works properly. Travis has been setup to check that
everything builds on stable and nightly and that there are no `clippy`
or `rustfmt` issues.