Document config->bar communication
To make updating the bar possible, it is required that the config can
communicate with the bar and push component updates. An outline of how
this is done has been added to the README.md.

Since this proposal includes setting a `text` for components in the
configuration file, the `text` field has been added to the struct and
tests have been changed accordingly.