Add config <-> component communication
This adds the first layer of communication required between the
configuration and the components.

This is implemented using an event loop which is notified by the
components whenever it needs to wake up. Once it has been notified, the
event loop queries the component responsible for the notification for
updates and propagates these updates if they changed the bar
configuration.

This also adds full documentation for all working features and provides
a little example which renders the state of the bar to the command line.

There currently is no way to notify the components that the frontend has
received an event which needs to be handled (click, hover, etc.).