Add component event notification
To allow for components to update themselves or execute arbitrary code
whenever they are clicked or other interacted with in any other way, it
is necessary to have a way to send events from the frontent to the bar
configuration.

This is done using the new `Bar::notify` method which needs to be called
whenever all components should be notified about a new event.

Since all events are global (positions are not component-relative), it
is necessary for each component to know about its own size. This is done
as a separate event called `PositionChange`. Currently it is necessary
for the frontend to keep all components updated using this event to make
sure the other position dependent events can be used properly.