Fix strfime format in clock component
The strftime format in the clock component was incorrectly specified as
"%H:%m", which will print the hours and month instead of the hours and
minutes.

To fix this the format has been changed to "%H:%M" which correctly
formats the time with zero-padded minutes.

Full documentation on strftime formats can be found here:
http://strftime.org/