Add dedicated border declaration to bar
Until now the intention of the `border` field in the component defaults
was to use the border in the defaults as a border for the complete bar.
However this is unintuitive and does not allow specifying a default
border for each component.

To make the usage of the general settings more intuitive and consistent,
a separate `border` field has been added to the general bar settings
which can be used to set the border for the complete bar. Then the
component settings in the defaults can be used as actual default values
for every component.