Rework deserialization and polling
Up until now the deserialization of components directly deserialized to
the type `Box<Component>`, where `Component` is the component trait.
This makes it hard to set the fallback values set in the `defaults`
field of the general bar settings.

To resolve this, the configuration is now deserialized into a
`Component` struct instead of the trait. Then after everything has been
deserialized it is converted to the runtime variant of `Box<Component>`.
This made it possible to add all the fallbacks and remove the `defaults`
field from the runtime state. The general bar settings now are also
separated, since the `Config` struct is only used as a wrapper for
initial deserialization now.

One problem with the public API was that the polling methods `recv` and
`try_recv` relied on `Arc<Mutex<Bar>>` to allow the consumer direct
access to the bar. This is blocking and requires handling of ugly locks
for the consumer.

To improve this, updating is now done in the main thread instead of the
event thread. The event thread just passes the component IDs to the main
thread and the main thread updates all components when `recv` or
`try_recv` are called. This removes all locks from the public API and
should make working with this crate a lot easier.

However this also introduced some slight drawbacks. Long update methods
might potentially delay both `recv` and `try_recv`. This also might
cause some significant delays until a component is actually updated,
since it relies on the user calling the `recv` and `try_recv` methods.
The best choice to prevent these drawbacks is to always drain all
updates using the `try_recv` method after initially polling with `recv`.
If this is done this should provide significant benefits over the
alternatives.